FIX: Widen unique-name counter to long so an Int32.MaxValue name suffix cannot overflow [UUM-145766]
Guard the collision counter against overflowing when the trailing digits equal Int32.MaxValue.

When an all-numeric base name parsed to exactly Int32.MaxValue, seeding the counter with trailingNumber + 1 overflowed int and the collision loop produced negative suffixes. The counter is now a long and the seeding addition is widened, so the suffix continues past Int32.MaxValue as a positive number.

Jira: https://jira.unity3d.com/browse/UUM-145766

+++ Packages/com.unity.inputsystem/InputSystem/Runtime/Utilities/StringHelpers.cs
@@ -328,7 +328,7 @@ namespace UnityEngine.InputSystem.Utilities
 
             var name = baseName;
             var nameIsUnique = false;
-            var namesTried = 1;
+            var namesTried = 1L;
 
             // If the name ends in digits, start counting from the given number.
             if (baseName.Length > 0)
@@ -339,7 +339,7 @@ namespace UnityEngine.InputSystem.Utilities
                 if (lastDigit != baseName.Length &&
                     int.TryParse(baseName.Substring(lastDigit), out var trailingNumber))
                 {
-                    namesTried = trailingNumber + 1;
+                    namesTried = (long)trailingNumber + 1;
                     baseName = baseName.Substring(0, lastDigit);
                 }
             }
